Change and document hypertree node actions.
Left clicking on a tab node switches to the tab. No need for displaying
the arrow now.

Tab nodes cannot be focused. Focusing was not producing a useful view
anyway. Other nodes can be focused on left clicking.

Actions are shown in the tooltips. Overall instructions are shown in the
visualization page.

No more animation of nodes on hover. Animation was causing problems with
centering and the subsequent layout turned out to be garbage. Removing
the animation at least keeps the important tree behaviour correct.